# openapi.yaml
openapi: 3.1.0
info:
  title: uma.moe API
  description: |
    Public API for uma.moe — Umamusume inheritance search, circle tracking, fan rankings, and trainer profiles.

    All endpoints listed here are public and require no authentication.
    If you have an API key, you can pass it via the `X-API-Key` header for usage tracking.
  version: 1.0.0
  contact:
    name: uma.moe
    url: https://uma.moe
servers:
  - url: https://uma.moe
    description: Production
...
components:
  schemas:
...
    UserFanRankingMonthly:
      type: object
      properties:
        viewer_id:
          type: integer
          format: int64
        trainer_name:
          type: string
          nullable: true
        year:
          type: integer
        month:
          type: integer
        total_fans:
          type: integer
          format: int64
        monthly_gain:
          type: integer
          format: int64
        active_days:
          type: integer
        avg_daily:
          type: number
          nullable: true
        avg_3d:
          type: number
          nullable: true
        avg_7d:
          type: number
          nullable: true
        avg_monthly:
          type: number
          nullable: true
        rank:
          type: integer
        circle_id:
          type: integer
          format: int64
          nullable: true
        circle_name:
          type: string
          nullable: true
        next_month_start:
          type: integer
          format: int64
          nullable: true
    MonthlyRankingsResponse:
      type: object
      properties:
        rankings:
          type: array
          items:
            $ref: "#/components/schemas/UserFanRankingMonthly"
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int64
        limit:
          type: integer
          format: int64
        total_pages:
          type: integer
          format: int64
        year:
          type: integer
        month:
          type: integer
    UserFanRankingAlltime:
      type: object
      properties:
        viewer_id:
          type: integer
          format: int64
        trainer_name:
          type: string
          nullable: true
        total_fans:
          type: integer
          format: int64
        total_gain:
          type: integer
          format: int64
        active_days:
          type: integer
        avg_day:
          type: number
          nullable: true
        avg_week:
          type: number
          nullable: true
        avg_month:
          type: number
          nullable: true
        rank:
          type: integer
        rank_total_fans:
          type: integer
        rank_total_gain:
          type: integer
        rank_avg_day:
          type: integer
        rank_avg_week:
          type: integer
        rank_avg_month:
          type: integer
        circle_id:
          type: integer
          format: int64
          nullable: true
        circle_name:
          type: string
          nullable: true
    AlltimeRankingsResponse:
      type: object
      properties:
        rankings:
          type: array
          items:
            $ref: "#/components/schemas/UserFanRankingAlltime"
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int64
        limit:
          type: integer
          format: int64
        total_pages:
          type: integer
          format: int64
    UserFanRankingGains:
      type: object
      properties:
        viewer_id:
          type: integer
          format: int64
        trainer_name:
          type: string
          nullable: true
        gain_3d:
          type: integer
          format: int64
        gain_7d:
          type: integer
          format: int64
        gain_30d:
          type: integer
          format: int64
        rank_3d:
          type: integer
        rank_7d:
          type: integer
        rank_30d:
          type: integer
        circle_id:
          type: integer
          format: int64
          nullable: true
        circle_name:
          type: string
          nullable: true
    GainsRankingsResponse:
      type: object
      properties:
        rankings:
          type: array
          items:
            $ref: "#/components/schemas/UserFanRankingGains"
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int64
        limit:
          type: integer
          format: int64
        total_pages:
          type: integer
          format: int64
        sort_by:
          type: string

    # ── Profile ────────────────────────────────────────────────
...
paths:
...
  /api/v4/rankings/monthly:
    get:
      tags: [Rankings]
      summary: Monthly fan gain rankings
      operationId: getMonthlyRankings
      parameters:
        - name: month
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 12
          description: Defaults to current month (JST)
        - name: year
          in: query
          schema:
            type: integer
          description: Defaults to current year (JST)
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 100
        - name: query
          in: query
          schema:
            type: string
          description: "viewer_id (exact), trainer_name or circle_name (partial)"
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [monthly_gain, total_fans, active_days, avg_daily, avg_3d, avg_7d, avg_monthly]
            default: monthly_gain
      responses:
        "200":
          description: Monthly rankings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyRankingsResponse"
  /api/v4/rankings/alltime:
    get:
      tags: [Rankings]
      summary: All-time fan rankings
      operationId: getAlltimeRankings
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 100
        - name: query
          in: query
          schema:
            type: string
          description: "viewer_id (exact), trainer_name or circle_name (partial)"
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [total_gain, total_fans, avg_day, avg_week, avg_month]
            default: total_gain
      responses:
        "200":
          description: All-time rankings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlltimeRankingsResponse"
  /api/v4/rankings/gains:
    get:
      tags: [Rankings]
      summary: Rolling gain rankings (3d/7d/30d)
      operationId: getGainsRankings
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 100
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [gain_3d, gain_7d, gain_30d]
            default: gain_30d
        - name: query
          in: query
          schema:
            type: string
          description: "viewer_id (exact), trainer_name or circle_name (partial)"
      responses:
        "200":
          description: Gains rankings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GainsRankingsResponse"

  # ── Profile ─────────────────────────────────────────────────
...